Split session handshake parts into focused modules
